requirements.txt
numpy==1.20.2
streamlit==0.82.0
tensorflow==2.3.0
pillow==8.2.0